Bug fixed of Gradual Scene Chagne Detection

+++ JM/lencod/inc/global.h
@@ -1539,10 +1539,10 @@ typedef struct video_par
   int AbrDet_Flag[300];
   int GraDet_Flag[300];
   float IPR[300];
-  int IPR1[300];
-  int SKP1[300];
-  int IPR2[300];
-  int SKP2[300];
+  float IPR1[300];
+  float SKP1[300];
+  float IPR2[300];
+  float SKP2[300];
 
   unsigned int PicWidthInMbs;
   unsigned int PicHeightInMapUnits;
